extract archive cache manager
